requirements.txt
# Core dependencies
streamlit>=1.32.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.18.0

# Google Cloud dependencies
google-cloud-storage>=3.0.0
google-cloud-firestore>=2.11.0
google-auth>=2.22.0

# PDF processing
llama-parse>=0.3.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Vector database
pinecone>=3.0.0

# OpenAI
openai>=1.12.0

# LlamaIndex and related packages
llama-index-core>=0.10.11
llama-index-embeddings-openai>=0.1.5
llama-index-llms-openai>=0.1.5
llama-index-postprocessor-flag-embedding-reranker>=0.1.1
llama-index-vector-stores-pinecone>=0.1.3
llama-index-readers-file>=0.1.3
llama-index-readers-llama-parse>=0.1.2
llama-cloud-services>=0.0.1
nest-asyncio>=1.5.0

# Hugging Face dependencies
llama-index-llms-huggingface>=0.1.3
llama-index-embeddings-huggingface>=0.1.3
transformers>=4.36.0
torch>=2.0.0
sentence-transformers>=2.2.2

# LangChain
langchain==0.3.19
langchain-openai==0.3.7
langchain-community==0.3.18
langchain-core==0.3.37
pydantic-settings==2.1.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
pytz>=2023.3
tenacity>=8.2.0

# Development tools
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0 